#36 Added importValue logic and related services/utils
